move reinitialize from components/ -> utils/
